created the backend for the login and register, now users cand login and register and their data will be stored in the database, modified the frontend to not be static anymore and also used JWT for security

+++ backend/src/main/java/com/thesis/backend/model/Section.java
@@ -17,7 +17,7 @@ import lombok.Setter;
 @Table(name="section")
 public class Section {
     @Id
-    private Integer sectionId;
+    private Long sectionId;
     private String title;
     private int sectionNumber;
 }
